Legal_Mcp/requirements.txt
Flask

python-dotenv

google-cloud-vision
google-generativeai

langchain
PyMuPDF
Pillow
pytest
pytest-mock
Werkzeug
Flask-Cors
gunicorn